move ChipAndCoreConstraint
