+++ pacman/model/constraints/__init__.py
@@ -14,6 +14,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from .abstract_constraint import AbstractConstraint
+from .chip_and_core_constraint import ChipAndCoreConstraint
 from .fixed_key_and_mask_constraint import FixedKeyAndMaskConstraint
 
-__all__ = ["AbstractConstraint", "FixedKeyAndMaskConstraint"]
+
+__all__ = ["AbstractConstraint", "ChipAndCoreConstraint",
+           "FixedKeyAndMaskConstraint"]
